hash-page-table: allocator restructure + 15-round convention sweep
Allocator restructure
- Each allocator owns its own OS pages; PageAllocator dependency dropped
  from Heap / Arena / Slab.
- Heap: hash-indexed pages[] + per-class warm list; xl-page descriptor
  table on its own mmap, page-rounded for grow/unmap symmetry.
- Typed-self everywhere: dispatch enters typed, vtable cast happens once
  at *Init compound literal, no `_dyn` thunks below the boundary.
- Extracted `_Os.{h,c}` shim (os_page_map / unmap / round_up) shared by
  every backend.
- Dirty-bit deep-validate gating on Page + Heap (LSB of `__magic`).
- New public accessors: HeapAllocator{PageCount,XlCount},
  PageAllocator{FootprintBytes,EntryCount}, BudgetAllocatorSlotCount.
- Arena.c: `zeroed` flag now honoured on in-chunk reuse (was returning
  stale tenant bytes after `ArenaAllocatorReset`).
- AllocatorStats accounting documented as typed-direct-skip (only the
  `Allocator *` dyn arm updates counters).

Bench / perf infrastructure
- Per-op + perf-record harnesses for HeapAllocator vs libc (Bench-Perf/).
- Cross-allocator comparison driver (Benchmark/) with libc / misra /
  misra-arena / misra-page split; Page footprint sampling via the new
  accessor.

Sys / parsers
- Sys/*: direct syscalls (Linux x86_64/aarch64 + Darwin BSD class) via
  Source/Misra/_Syscall.h; FEATURE_DIRECT_SYSCALL gates the wrapper
  path. ErrnoOf(ret) unifies `<0`-means-`-errno` (direct) and `-1`+errno
  (libc fallback).
- Proc.c: child exit path on execve failure (was returning into parent
  code as a second process); wait4 sentinel switched to `< 0`;
  GetCurrentExecutablePath readlink sentinel switched to `< 0`;
  stdin/stdout/stderr pipe initialisers fixed (was leaving fd 0).
- Socket.c / Dns.c / Dir.c / Proc.c: alias wrappers purged; canonical
  `LOG_SYS_ERROR(ErrnoOf(ret), ...)` everywhere.
- Mutex.h+c: gating unified on FEATURE_DIRECT_SYSCALL; dead pthread
  fallback removed; NULL guards added to Lock/Unlock.
- Backtrace / SymbolResolver / PdbCache / MachoCache stay outside the
  Sys umbrella (parser-name collision risk); umbrella comment lists the
  full carve-out.
- Parsers: ElfFindSection / PeFindSection / MachoFindSection now take
  Str / Zstr / char* via _Generic; dns_resolve_4_one same.
- Elf.c: `class` field renamed to `elf_class` (C++ keyword).
- Dns.c: VecPushBackR failure path no longer leaks the Str name.

Containers
- Validator-before-field-access in Map.c / List.c / Graph.c
  validate_*; closes NULL-deref on a magic-OK handle with NULL
  allocator pointer.
- Float / Int observers const-cascaded across Memory / Convert / Compare
  / Math / Private headers (and their callers).
- Alias wrappers removed: MapTryGetPtr, MapGetOrInsertPtr,
  BitVecInitWithCapacityMacro. Callers migrated.
- Iter / StrIter: direction-aware peek (PeekNext / PeekPrev) wired
  through the `dir` field; macro args parenthesised.

Std / freestanding
- Io.c read-text-int family (_read_u8..u64 / i8..i64) consolidated via
  `_MAKE_READ_TXT_INT` stamping macro (net -460 LOC).
- _Freestanding.c: `_setjmp` deduplicated as `__attribute__((alias(
  "setjmp")))`; jmp_buf size docs corrected (x86_64 64 B, aarch64 104 B).
- Zstr.c parseInt: removed `s[-1]` UB when no digits consumed.
- Types.h: NULL is `((void *)0)`; FORMAT_STRING uses `__attribute__`
  (was `__attribute`); INVERT_ENDIANNESS{4,8} body and arg paren-fixed
  (was silently corrupting the high half for any direct caller).

Tests
- TestRunner: `expect_failure=false` now installs OnAbort and routes
  through one setjmp before the test body (previously skipped the
  setjmp install AND the OnAbort reset).
- Tests/Std: `&alloc.base` replaced with `ALLOCATOR_OF(&alloc)`;
  white-box private-field reads/writes carry `// intentional bypass:`
  comments per the new CONVENTIONS phrase.
- Tests/meson.build: PdbCache + MachoCache gated on `opt_file
  and opt_sys_dir` so minimal builds link.
- Fuzz/Harness/Str.{c,h}: deleted (stub: referenced non-existent APIs).

Docs / build
- CODING-CONVENTIONS.md: new sections for `intentional bypass:`
  canonical phrase, AllocatorStats typed-direct-skip, .c-local
  macro-stamping pattern as approved alternative to alias wrappers;
  rule reaffirmed: NO MISRA_ prefix on public macros (only include
  guards + MISRA_SYS_* constants); NO alias wrappers (neither
  `#define foo bar` nor `static inline T foo(...){return bar(...);}`);
  _Generic inlined at each call site, Zstr arms paired with char*.
- README.md: minimal-build example with full disable-flag list and
  cascade explanation (alloc_debug auto-enables map + sys_backtrace +
  parser_elf + parser_dwarf); file / iter / alloc_arena flagged as
  not-currently-disable-able with rationale.
- Misra.h / Std.h / Sys.h umbrellas: FEATURE_* gates and carve-outs
  brought in line with the actual feature matrix; Container.h umbrella
  now exposes Buf.h.
- meson.build: dead libm dep removed.
- File-path header comments normalised to lowercase
  `/// file: relative/path.ext` across every header and source.

Build / test status: Linux 104/104 + Mac 99/99 green at this tip.

+++ Source/Misra/Parsers/KvConfig.c
@@ -1,8 +1,14 @@
+/// file      : parsers/kvconfig.c
+/// author    : Siddharth Mishra (admin@brightprogrammer.in)
+/// This is free and unencumbered software released into the public domain.
+///
+/// Key-value configuration parser: `key = value` lines into a Map(Str, Str).
+
 #include <Misra/Parsers/KvConfig.h>
-#include <Misra/Std/Zstr.h>
 #include <Misra/Std/Container/Map/Private.h>
-#include <Misra/Std/Memory.h>
 #include <Misra/Std/Log.h>
+#include <Misra/Std/Memory.h>
+#include <Misra/Std/Zstr.h>
 
 
 static bool kvconfig_is_space(char c) {
@@ -37,7 +43,7 @@ static bool kvconfig_parse_bool_value(const Str *value, bool *out) {
 
 static bool kvconfig_parse_i64_value(const Str *value, i64 *out) {
     Zstr endptr = NULL;
-    long long parsed;
+    i64  parsed;
 
     if (!out) {
         LOG_FATAL("Expected valid integer output pointer");
@@ -49,13 +55,13 @@ static bool kvconfig_parse_i64_value(const Str *value, i64 *out) {
         return false;
     }
 
-    *out = (i64)parsed;
+    *out = parsed;
     return true;
 }
 
 static bool kvconfig_parse_f64_value(const Str *value, f64 *out) {
     Zstr endptr = NULL;
-    double parsed;
+    f64  parsed;
 
     if (!out) {
         LOG_FATAL("Expected valid float output pointer");
